requirements.txt
beautifulsoup4==4.11.1
Flask==2.1.3
numpy==1.21.5
pandas==1.4.3
Pillow==9.3.0
requests==2.28.1
scikit_learn==1.1.3
gunicorn==20.0.4
regex==2020.2.20
IMDbPY==6.8